py bindings -- add more usage help and unbiased timings to autotuner

+++ py_bindings/autotune.py
@@ -96,6 +96,8 @@ import shutil
 import autotune_template
 import psutil
 import operator
+import glob
+import re
 from valid_schedules import *
 
 sys.path += ['../util']
@@ -157,7 +159,8 @@ class AutotuneParams:
       -runner_file             s Runner C++ filename, defaults to within runner/ directory.
       -summary_file            s Summary output filename, defaults to summary.txt
       -plot_file               s Convergence plot filename, defaults to plot.png
-
+      -unbiased_file           s Stores unbiased timing comparisons, defaults to unbiased.txt (summary has biased times)
+      
     Experimental Features:
     
       -max_nontrivial          n When generating random schedules, max number of nontrivial (non-root/inline) funcs
@@ -236,7 +239,8 @@ class AutotuneParams:
     runner_file = 'default_runner.cpp'
     summary_file = 'summary.txt'
     plot_file = 'plot.png'
-
+    unbiased_file = 'unbiased.txt'  # For unbiased timing comparisons (summary is biased)
+    
     def __init__(self, argd={}):
         for (key, value) in argd.items():
             if not hasattr(self, key):
@@ -847,8 +851,11 @@ def run_limit(L, timeout, last_line=False, time_from_subproc=False, shell=False,
         ans = ans.strip().split('\n')[-1].strip()
     return proc.returncode, ans
 
+def identity_prefix():
+    return 'f'
+    
 def schedule_ref_output(p, schedule, j):
-    return os.path.join(p.tune_dir, 'f' + schedule.identity() + '_%d'%j + p.image_ext)
+    return os.path.join(p.tune_dir, identity_prefix() + schedule.identity() + '_%d'%j + p.image_ext)
     
 def default_tester(input, out_func, p, filter_func_name, allow_cache=True):
     cache = {}
@@ -884,7 +891,7 @@ def default_tester(input, out_func, p, filter_func_name, allow_cache=True):
                     trials = trials_override[i]
                 else:
                     trials = p.trials
-            binary_file = os.path.join(p.tune_dir, 'f' + schedule.identity())
+            binary_file = os.path.join(p.tune_dir, identity_prefix() + schedule.identity())
             mode_str = 'compile' if compile else 'run'
             
             save_filename = ''
@@ -1091,7 +1098,7 @@ def autotune(filter_func_name, p, tester=default_tester, constraints=Constraints
 
     nref = 0
     for (ref_name, ref_schedule_str) in scope.get('tune_ref_schedules', {}).items():
-        currentL.append(constraints.constrain(Schedule.fromstring(out_func, ref_schedule_str, ref_name, 0, len(currentL))))
+        currentL.append(constraints.constrain(Schedule.fromstring(out_func, ref_schedule_str, 'ref_' + ref_name, 0, len(currentL))))
         nref += 1
 
         #print 'seed_schedule new_vars', seed_schedule.new_vars()
@@ -1168,6 +1175,7 @@ def autotune(filter_func_name, p, tester=default_tester, constraints=Constraints
         sys.stdout.flush()
         #autotune_plot.main((os.path.join(p.tune_dir, p.summary_file), os.path.join(p.tune_dir, p.plot_file)))
         os.system('python autotune_plot.py "%s" "%s"' % (os.path.join(p.tune_dir, p.summary_file), os.path.join(p.tune_dir, p.plot_file)))
+        os.system('python autotune.py time "%s" "%s"' % (p.tune_dir, os.path.join(p.tune_dir, p.unbiased_file)))
 
 import inspect
 _scriptfile = inspect.getfile(inspect.currentframe()) # script filename (usually with path)
@@ -1326,16 +1334,33 @@ def test():
     import autotune_test
     autotune_test.test()
     test_valid_schedules()
+
+def split_doublequote(s):
+    L = re.compile(r'''((?:"[^"]*")+)''').split(s)
+    ans = []
+    for x in L:
+        if x.startswith('"'):
+            ans.append(x)
+        else:
+            ans.extend(x.split())
+    return ans
     
 def main():
     (args, argd) = parse_args()
     all_examples = 'blur dilate boxblur_cumsum boxblur_sat erode snake interpolate bilateral_grid camera_pipe local_laplacian'.split() # local_laplacian'.split()
     if len(args) == 0:
-        print 'autotune test|print|autotune examplename.filter_func|test_sched|test_fromstring|test_variations'
+        print 'autotune test|print|test_sched|test_fromstring|test_variations'
         print '  Internal use.'
         print
-        print 'autotune example [%s|all|list.txt]'%('|'.join(all_examples))
-        print '  Run on one of the examples (or all examples, or a whitespace separated list in a file with .txt extension)'
+        print 'autotune X.Y.filter_func [switches]'
+        print '  Direct use of autotuner: "from X.Y import filter_func" should give a function a la examples/*.py'
+        print
+        print 'autotune example [%s|all|list.txt] [switches]' %('|'.join(all_examples))
+        print '  Helper to tune one example (or all examples, or a whitespace separated list in a file with .txt extension).'
+        print '  Supplies reasonable arguments for the example.'
+        print
+        print 'autotune time tune_dir [log_timefile.txt]'
+        print '  Tuner timings are biased. Run this to get unbiased comparison of ref schedules against best tuned schedule.'
         print '\n'.join(x[4:] for x in AutotuneParams.__doc__.split('\n'))
         sys.exit(0)
     if args[0] == 'test':
@@ -1351,8 +1376,26 @@ def main():
         exampleL = all_examples if examplename == 'all' else [examplename]
         if examplename.lower().endswith('.txt'):
             exampleL = open(examplename,'rt').read().strip().split()
+            
+        def get_tune_dir(examplename):
+            return 'tune_%s'%examplename
+
+        existL = []
         for examplename in exampleL:
-            tune_dir = 'tune_%s'%examplename
+            tune_dir = get_tune_dir(examplename)
+            if os.path.exists(tune_dir):
+                existL.append(tune_dir)
+        if len(existL):
+            print 'The following directories exist, remove them?'
+            print
+            print '\n'.join('  ' + x for x in existL)
+            print 
+            ok = raw_input('Remove [y/n]? ')
+            if ok.strip().lower() != 'y':
+                sys.exit(1)
+            
+        for examplename in exampleL:
+            tune_dir = get_tune_dir(examplename)
             if os.path.exists(tune_dir):
                 shutil.rmtree(tune_dir)
             if examplename == 'local_laplacian':
@@ -1365,6 +1408,79 @@ def main():
             elif examplename in ['camera_pipe', 'bilateral_grid']:
                 rest = '-generations 150'.split() + rest
             system('python autotune.py autotune examples.%s.filter_func -tune_dir "%s" %s' % (examplename, tune_dir, ' '.join(rest)))
+    elif args[0] == 'time':
+        if len(args) < 2:
+            print >> sys.stderr, 'Expected 2 arguments'
+            sys.exit(1)
+        tune_dir = args[1]
+        log_timefile = None
+        if len(args) > 2:
+            log_timefile = args[2]
+        p = AutotuneParams(argd)
+        summary = open(os.path.join(tune_dir, p.summary_file), 'rt').read().split('\n')
+        reset = False
+        tval = None
+        indiv = None
+        for line in summary:
+            line = line.strip()
+            if line.startswith('Generation'):
+                pass #print '*', line
+            elif line == '-'*len(line) and len(line) >= 1:
+                #print 'turned reset on'
+                reset = True
+            elif reset:
+                if len(line.split()) >= 3:
+                    #print line.split()
+                    tval = line.split()[0]
+                    indiv = identity_prefix() + line.split()[1]
+                    #print 'set', tval, indiv
+                #print 'turned reset off'
+                reset = False
+        if indiv is None:
+            print 'Did not find best individual, timing failed'
+            sys.exit(1)
+        verbose = False
+        if verbose:
+            print 'Found best      individual', indiv, '(biased time previously reported: %s)'%tval
+        refL = sorted(glob.glob(os.path.join(os.path.abspath(tune_dir), 'f000*compile.sh')))
+        if len(refL) == 0:
+            print 'Could not find reference schedules, timing failed'
+            sys.exit(1)
+        ref_indiv = os.path.split(refL[-1])[-1].replace('_compile.sh', '')
+        if verbose:
+            print 'Found reference individual', ref_indiv
+            print
+
+        def do_time(indiv, ref_indiv):
+            sh = open(os.path.join(tune_dir, indiv + '_compile.sh'), 'rt').read().strip()
+            #print sh
+            #L = shlex.split(sh.replace('"', '\x255'))
+            L = split_doublequote(sh)
+            #print L
+            #sys.exit(1)
+            L_orig = L[10].strip('"')
+            L[8] = os.path.abspath(os.path.join(tune_dir, indiv))
+            if indiv != ref_indiv:
+                L[10] = '"' + os.path.abspath(os.path.join(tune_dir, os.path.split(L_orig)[-1])) + '"'
+            sh = ' '.join(L)
+            if verbose:
+                print sh
+            #print
+            subprocess.check_output(sh, shell=True)
+            output = subprocess.check_output(sh.replace('autotune_compile_child', 'autotune_run_child'), shell=True)
+            return output.strip().split('\n')[-1]
+            
+        out1 = do_time(indiv, ref_indiv)
+        out2 = do_time(ref_indiv, ref_indiv)
+        sout = 'Unbiased Timings:\n'
+        sout += 'best      ' + indiv + ' ' + out1 + '\n'
+        sout += 'reference ' + ref_indiv + ' ' + out2 + '\n'
+        sout += '\n'
+        print sout,
+        if log_timefile is not None:
+            f = open(log_timefile, 'at')
+            f.write(sout)
+            f.close()
     elif args[0] == 'test_random':
         import autotune_test
         global use_random_blocksize
